@@ -1,9 +1,10 @@
 {
     "name": "@semivan/eslint-config",
-    "version": "1.2.8",
+    "version": "2.0.0",
     "description": "JavaScript ESLint config",
     "author": "Ivan Semenikhin <semenihin.ivan@gmail.com>",
     "license": "MIT",
+    "type": "module",
     "exports": {
         ".": "./src/index.js",
         "./rules": "./src/rules/index.js",
@@ -17,8 +18,9 @@
         "url": "https://github.com/semivan/javascript"
     },
     "devDependencies": {
-        "@stylistic/eslint-plugin": "^1.7.0",
-        "eslint": "^8.57.0"
+        "@stylistic/eslint-plugin": "^2.12.0",
+        "eslint": "^9.16.0",
+        "globals": "^15.13.0"
     },
     "peerDependencies": {
         "@stylistic/eslint-plugin": "*",
