@@ -1,14 +1,18 @@
-const jsRules = require('../js');
+import jsRules from '../js/index.js';
 
-module.exports = {
+export default {
     'dot-notation': 'off',
     'no-implied-eval': 'off',
     'no-throw-literal': 'off',
     'prefer-destructuring': 'off',
+    'prefer-promise-reject-errors': 'off',
     'require-await': 'off',
 
     // Disallow awaiting a value that is not a Thenable
     '@typescript-eslint/await-thenable': 'error',
+
+    // Require return statements to either always or never specify values
+    // '@typescript-eslint/consistent-return': 'off',
 
     // Enforce consistent usage of type exports
     '@typescript-eslint/consistent-type-exports': ['error', {
@@ -22,6 +26,9 @@ module.exports = {
     // Enforce naming conventions for everything across a codebase
     '@typescript-eslint/naming-convention': 'off',
 
+    // Disallow using the delete operator on array values
+    '@typescript-eslint/no-array-delete': 'error',
+
     // Require .toString() to only be called on objects which provide useful information when stringified
     '@typescript-eslint/no-base-to-string': 'error',
 
@@ -29,6 +36,9 @@ module.exports = {
     '@typescript-eslint/no-confusing-void-expression': ['error', {
         ignoreVoidOperator: true,
     }],
+
+    // Disallow using code marked as @deprecated
+    '@typescript-eslint/no-deprecated': 'error',
 
     // Disallow duplicate constituents of union or intersection types
     '@typescript-eslint/no-duplicate-type-constituents': ['error', {
@@ -64,7 +74,7 @@ module.exports = {
     '@typescript-eslint/no-redundant-type-constituents': 'error',
 
     // Disallow throwing literals as exceptions
-    '@typescript-eslint/no-throw-literal': jsRules['no-throw-literal'] ?? 'off',
+    // '@typescript-eslint/no-throw-literal': jsRules['no-throw-literal'] ?? 'off',
 
     // Disallow unnecessary equality comparisons against boolean literals
     '@typescript-eslint/no-unnecessary-boolean-literal-compare': 'error',
@@ -75,11 +85,17 @@ module.exports = {
     // Disallow unnecessary namespace qualifiers
     '@typescript-eslint/no-unnecessary-qualifier': 'error',
 
+    // Disallow unnecessary template expressions
+    '@typescript-eslint/no-unnecessary-template-expression': 'error',
+
     // Disallow type arguments that are equal to the default
     '@typescript-eslint/no-unnecessary-type-arguments': 'error',
 
     // Disallow type assertions that do not change the type of an expression
     '@typescript-eslint/no-unnecessary-type-assertion': 'error',
+
+    // Disallow type parameters that aren't used multiple times
+    '@typescript-eslint/no-unnecessary-type-parameters': 'error',
 
     // Disallow calling a function with a value with type any
     '@typescript-eslint/no-unsafe-argument': 'error',
@@ -99,6 +115,9 @@ module.exports = {
     // Disallow returning a value with type any from a function
     '@typescript-eslint/no-unsafe-return': 'error',
 
+    // Disallow type assertions that narrow a type
+    '@typescript-eslint/no-unsafe-type-assertion': 'error',
+
     // Require unary negation to take a number
     '@typescript-eslint/no-unsafe-unary-minus': 'off',
 
@@ -108,8 +127,14 @@ module.exports = {
     // Enforce non-null assertions over explicit type casts
     '@typescript-eslint/non-nullable-type-assertion-style': 'off',
 
+    // Disallow throwing non-Error values as exceptions
+    '@typescript-eslint/only-throw-error': jsRules['no-throw-literal'] ?? 'off',
+
     // Require destructuring from arrays and/or objects
     '@typescript-eslint/prefer-destructuring': jsRules['prefer-destructuring'] ?? 'off',
+
+    // Enforce the use of Array.prototype.find() over Array.prototype.filter() followed by [0] when looking for a single result
+    '@typescript-eslint/prefer-find': 'error',
 
     // Enforce includes method over indexOf method
     '@typescript-eslint/prefer-includes': 'error',
@@ -120,6 +145,12 @@ module.exports = {
     // Enforce using concise optional chain expressions instead of chained logical ands,
     // negated logical ors, or empty objects
     '@typescript-eslint/prefer-optional-chain': 'error',
+
+    // Require using Error objects as Promise rejection reasons
+    '@typescript-eslint/prefer-promise-reject-errors': ['error', {
+        allowThrowingAny: false,
+        allowThrowingUnknown: false,
+    }],
 
     // Require private members to be marked as readonly if they're never modified outside of the constructor
     '@typescript-eslint/prefer-readonly': 'off',
@@ -141,6 +172,9 @@ module.exports = {
 
     // Require any function or method that returns a Promise to be marked async
     '@typescript-eslint/promise-function-async': 'off',
+
+    // Enforce that get() types should be assignable to their equivalent set() type
+    '@typescript-eslint/related-getter-setter-pairs': 'off',
 
     // Require Array#sort and Array#toSorted calls to always provide a compareFunction
     '@typescript-eslint/require-array-sort-compare': 'off',
@@ -171,4 +205,7 @@ module.exports = {
 
     // Enforce unbound methods are called with their expected scope
     '@typescript-eslint/unbound-method': 'error',
+
+    // Enforce typing arguments in Promise rejection callbacks as unknown
+    '@typescript-eslint/use-unknown-in-catch-callback-variable': 'off',
 };
