@@ -1,10 +1,13 @@
-module.exports = {
-    ...require('./best-practices.js'),
-    ...require('./errors.js'),
-    ...require('./es6.js'),
-    ...require('./node.js'),
-    ...require('./strict.js'),
-    ...require('./style.js'),
-    ...require('./stylistic.js'),
-    ...require('./variables.js'),
+import formatting from './formatting.js';
+import possibleProblems from './possible-problems.js';
+import stylistic from './stylistic.js';
+import suggestions from './suggestions.js';
+
+// Object.values(document.querySelectorAll('.rule__name')).map(code => `'${code.innerText}': null`).join(', ')
+
+export default {
+    ...formatting,
+    ...possibleProblems,
+    ...stylistic,
+    ...suggestions,
 };
